+++ src/main.ts
@@ -5,4 +5,4 @@ import router from './router/index'
 
 const app = createApp(App)
 
-createApp(App).use(router).mount('#app')
+app.use(router).mount('#app')