+++ .cz-config.js
@@ -18,6 +18,7 @@ module.exports = {
     { name: 'components' },
     { name: 'package' },
     { name: 'request' },
+    { name: 'utils' },
     { name: 'hook' }
   ],
   // it needs to match the value for field type. Eg.: 'fix'