+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - ✨ 新增(router): 引入 vue-router ([c3ec5b6](https://github.com/2401345934/webpack-vue-demo/commit/c3ec5b6))
 - ✨ 新增(webpack): 修改 webpack 配置 优化打包性能 ([be1cf08](https://github.com/2401345934/webpack-vue-demo/commit/be1cf08))
 - ✨ 新增(webpack): 增加自动解析 webpack ([07f4ba6](https://github.com/2401345934/webpack-vue-demo/commit/07f4ba6))
+- ✨ 新增(webpack): webpack 修改配置优化打包速度 ([0ef33b1](https://github.com/2401345934/webpack-vue-demo/commit/0ef33b1))
 - 初始化环境 ([f5d5e39](https://github.com/2401345934/webpack-vue-demo/commit/f5d5e39))
 - 推送 ([ee2f81e](https://github.com/2401345934/webpack-vue-demo/commit/ee2f81e))
 - 引入 babel ([f3917aa](https://github.com/2401345934/webpack-vue-demo/commit/f3917aa))